fixed zip writing for PPMD
